+++ plugin/plugin.interface.ts
@@ -1,0 +1,6 @@
+import { GlobalStore } from '../store/global.store';
+import { Session } from '../store/session';
+
+export interface PluginInterface {
+  evaluate(): Promise<void>;
+}