+++ run.sh
@@ -1,6 +1,3 @@
 #!/bin/bash
 
-# FIXME: nvm is not detected
-# nvm use
-
 PORT=3142 npm run dev